+++ arch/sim/src/sim/up_unblocktask.c
@@ -135,13 +135,7 @@ void up_unblock_task(FAR struct tcb_s *tcb)
            * interrupt driven environment.
            */
 
-          rtcb = this_task();
-          if (rtcb->xcp.sigdeliver)
-            {
-              sinfo("Delivering signals TCB=%p\n", rtcb);
-              ((sig_deliver_t)rtcb->xcp.sigdeliver)(rtcb);
-              rtcb->xcp.sigdeliver = NULL;
-            }
+          sim_sigdeliver();
         }
     }
 }
